fix bug allowing uncaptured builds
This commit fixes a bug that permitted a build to remain uncaptured at the end
of a round. validateAction rejects building Actions by players who don't hold
any cards that could capture the build, but it erroneously allowed a player to
use their last such card to capture other piles without also capturing the
build. (Because it is clearly contrary to the spirit of the game, and also
almost always sub-optimal in practice, some statements of the rules do not
explicitly forbid this, seeming to consider it implicitly understood to be
impermissible.)

This commit causes validateAction to reject a capturing Action by a player who
controls an outstanding build if no card that could capture it would remain in
their hand. Along with the existing check at build creation and the restriction
against trailing with a controlled build, this is sufficient to enforce the
rule that a build must be captured in the round in which it is created.